+++ internal/testutil/harness/doc.go
@@ -1,4 +1,4 @@
-// Package e2e provides end-to-end testing utilities and fixtures.
+// Package harness provides a shared test harness for BDD and integration tests.
 //
 // # Overview
 //
@@ -6,4 +6,4 @@
 // - E2E test fixtures
 // - Test helpers for integration testing
 // - Scenario builders for complex test cases
-package e2e
+package harness
